Time: 211 ms (10.25%) | Memory: 88.1 MB (15.23%) - LeetSync

+++ 2188-minimized-maximum-of-products-distributed-to-any-store/minimized-maximum-of-products-distributed-to-any-store.cpp
@@ -1,10 +1,8 @@
 class Solution {
 public:
     int helper(vector<int>& q, int n, int m){
-        // long long x = accumulate(q.begin(), q.end(), 0);
-        // n -= ceil((double)x/m);
         for(auto x: q){
-            n -= ceil((double)x/m);
+            n -= ceil((float)x/m);
         }
 
         return n >= 0;
